LineChart->MetricLegendThreshold
