+++ packages/ui-components/src/components/LineIntervalChart/LineIntervalChart.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { ScaleLinear, ScaleTime } from "d3-scale";
 import { Group } from "@visx/group";
 import { Timeseries } from "@actnowcoalition/metrics";
-
 import { LineChart } from "../LineChart";
 import { RectClipGroup } from "../RectClipGroup";
