+++ packages/ui-components/src/components/Markdown/Markdown.tsx
@@ -2,9 +2,8 @@ import * as React from "react";
 import { ReactMarkdownOptions } from "react-markdown/lib/react-markdown";
 import { MarkdownBody } from "./Markdown.style";
 
-export const Markdown: React.FC<ReactMarkdownOptions> = ({
-  children,
-  ...otherProps
-}) => <MarkdownBody {...otherProps}>{children}</MarkdownBody>;
+export const Markdown = ({ children, ...otherProps }: ReactMarkdownOptions) => (
+  <MarkdownBody {...otherProps}>{children}</MarkdownBody>
+);
 
 export default Markdown;
