+++ packages/ui-components/src/components/MetricCatalogContext/MetricCatalogContext.tsx
@@ -48,10 +48,16 @@ const MetricCatalogContext = createContext<MetricCatalog>(defaultMetricCatalog);
  * @param children The component tree that we want to have access to the metric catalog.
  * @returns React.ContextProvider with the given metricCatalog
  */
-export const MetricCatalogProvider: React.FC<{
+
+export interface MetricCatalogProviderProps {
   metricCatalog: MetricCatalog;
   children: React.ReactNode;
-}> = ({ metricCatalog, children }) => (
+}
+
+export const MetricCatalogProvider = ({
+  metricCatalog,
+  children,
+}: MetricCatalogProviderProps) => (
   <MetricCatalogContext.Provider value={metricCatalog}>
     {children}
   </MetricCatalogContext.Provider>
