+++ packages/ui-components/src/components/MetricCompareTable/index.ts
@@ -1,2 +1,1 @@
 export * from "./MetricCompareTable";
-export * from "./interfaces";
