+++ packages/ui-components/src/components/MetricLegendThreshold/MetricLegendThreshold.tsx
@@ -2,8 +2,45 @@ import React from "react";
 import { Stack, Typography, Box } from "@mui/material";
 import { LegendThreshold } from "../LegendThreshold";
 import { useMetricCatalog } from "../MetricCatalogContext";
-import { CategoryItem, MetricLegendThresholdProps } from "./interfaces";
 import { getMetricCategoryItems } from "./utils";
+import { Metric } from "@actnowcoalition/metrics";
+
+export interface MetricLegendThresholdProps {
+  /** Orientation of the bars. */
+  orientation: "horizontal" | "vertical";
+  /** Metric to display thresholds for. */
+  metric: Metric | string;
+  /** Whether to show category labels. Does not affect start/endLabels */
+  showLabels?: boolean;
+  /** Optional label for the left or top side of the thermometer. */
+  startLabel?: React.ReactNode;
+  /** Optional label for the right or bottom side of the thermometer. */
+  endLabel?: React.ReactNode;
+  /** Whether or not to display metric name and info. If false, only thermometer is displayed. */
+  includeOverview?: boolean;
+  /** Border radius of the bars */
+  borderRadius?: number;
+  /** Width of the bars */
+  width?: number;
+  /** Height of the bars */
+  height?: number;
+  /** Optional other props. */
+  otherSvgProps?: Omit<
+    React.SVGProps<SVGSVGElement>,
+    keyof MetricLegendThresholdProps
+  >;
+}
+
+export interface CategoryItem {
+  /** Category name (e.g. "High") */
+  name: string;
+  /** Category color */
+  color: string;
+  /** Description of the category */
+  description: string | undefined;
+  /** Formatted value of the threshold at the end of the current category */
+  endThreshold?: string;
+}
 
 const getItemColor = (item: CategoryItem) => item.color;
 const getItemLabelHorizontal = (item: CategoryItem) => item.endThreshold ?? "";
